fix: tag expenses report and budget editors not reloading, breaking tour cy
